services/ai-microservice-engine/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-dotenv>=0.19.0
requests>=2.26.0
openai>=0.27.0
redis>=4.3.4
psycopg2-binary>=2.9.3
pydantic>=1.8.2
sqlalchemy>=1.4.23
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
aiohttp>=3.8.1
tenacity>=8.0.1
numpy>=1.21.2
pandas>=1.3.3
scikit-learn>=0.24.2
